requirements.txt
result
discord.py
